pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "csv-analyzer"
version = "0.1.0"
description = "A Python with pandas for analyze CSV file into data"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "requests>=2.31.0",
  "pydantic==2.12.5",
  "PyYAML==6.0.3",
  "setuptools==80.10.2",
  "pytest==9.0.2",
  "pandas==3.0.1",
]

[tool.setuptools.packages.find]
where = ["program"]